Dielectric material, bug fix for intersection, show render in window, print timings to std out

+++ src/defs.rs
@@ -1,2 +1,2 @@
-//pub type Real = f64;
-pub type Real = f32;
+pub type Real = f64;
+//pub type Real = f32;
